"Cell2" -> "Cell", finally
